+++ desmost/src/decompiler/decompile.ts
@@ -72,9 +72,14 @@ export function desmos_to_ast(
  */
 export function ast_to_source(ast: SemiStructuredAst, options: DesmostOptions): string
 {
+	let globals = ast.globals.map(inc => emit_incantation(inc, options, { expand: true })).join("\n");
+
 	return (
-        ast.globals.map(inc => emit_incantation(inc, options, { expand: true })).join("\n")
-      + (options.keep_leading_blanks ? "\n" : "\n\n")
+      globals
+      + (
+			globals.length === 0 ? ""
+			: options.keep_leading_blanks ? "\n" : "\n\n"
+		)
       + ast.locals.map(expr => emit_expression(expr, options)).join("\n")
 	);
 }