+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 <br>
 
 
+## v0.11.3
+
+### Fixes
+- 【compiler】Fix `/dark`
+
+
 ## v0.11.2
 
 ### New